backend/requirements.txt
#requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
gunicorn==21.2.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.0
email-validator==2.1.0
argon2-cffi==23.1.0
